Sagas and reducers update
