+++ src/store/mutations.js
@@ -1,20 +1,25 @@
 export const REQUEST_CHANGE_USERNAME = `REQUEST_CHANGE_USERNAME`;
 export const CHANGE_USERNAME = `CHANGE_USERNAME`;
+
 export const SET_STATE = `SET_STATE`;
 export const REQUEST_CLEAR_STATE = `REQUEST_CLEAR_STATE`;
+
 export const REQUEST_USER = `REQUEST_USER`;
 export const FAILED_SET_USER = `FAILED_SET_USER`;
+
 export const REQUEST_UPDATE_FAV_DRINKS = `REQUEST_UPDATE_FAV_DRINKS`;
 export const FAILED_UPDATE_FAV_DRINKS = `FAILED_UPDATE_FAV_DRINKS`;
+
 export const REQUEST_UPDATE_WHO_CURRENT_USER_FOLLOWS = `REQUEST_UPDATE_WHO_CURRENT_USER_FOLLOWS`;
 export const SUCCESSFUL_UPDATE_WHO_CURRENT_USER_FOLLOWS = `SUCCESSFUL_UPDATE_WHO_CURRENT_USER_FOLLOWS`;
 export const FAILED_UPDATE_WHO_CURRENT_USER_FOLLOWS = `FAILED_UPDATE_WHO_CURRENT_USER_FOLLOWS`;
 export const ADD_USER_TO_FOLLOWING = `ADD_USER_TO_FOLLOWING`;
 export const REMOVE_USER_FROM_FOLLOWING = `REMOVE_USER_FROM_FOLLOWING`;
+
 export const REQUEST_UPDATE_AVI =  `REQUEST_UPDATE_AVI`;
 export const SUCCESSFUL_UPDATE_AVI = `SUCCESSFUL_UPDATE_AVI`;
 export const FAILED_UPDATE_AVI = `FAILED_UPDATE_AVI`;
-export const CLEAR_AVI = `CLEAR_AVI`;
+
 export const REGISTER = `REGISTER`;
 export const REGISTER_SUCCESS = `REGISTER_SUCCESS`;
 export const REGISTER_FAILED = `REGISTER_FAILED`;
@@ -83,10 +88,6 @@ export const successfulUpdateAvi = (state) => ({
 
 export const failedUpdateAvi = () => ({
   type: FAILED_UPDATE_AVI
-});
-
-export const clearAvi = () => ({
-  type : CLEAR_AVI
 });
 
 export const requestRegistration = (registration) => ({
